Replace numpy.float with float because of the following deprecation warning:
DeprecationWarning: `np.float` is a deprecated alias for the builtin `float`. To silence this warning, use `float` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.float64` here.
 Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
dcalib = numpy.array([1, 1.3, 2, 3.5, 4, 5, 1.3, 6, 9.1], dtype=numpy.float)

+++ src/odemis/acq/stitching/_weaver.py
@@ -182,7 +182,7 @@ class Weaver(metaclass=ABCMeta):
         # To avoid overflows, we need to clip the results to the dtype range.
         if numpy.issubdtype(tile.dtype, numpy.integer):
             maxval = numpy.iinfo(tile.dtype).max
-        elif numpy.issubdtype(tile.dtype, numpy.float):
+        elif numpy.issubdtype(tile.dtype, float):
             maxval = numpy.finfo(tile.dtype).max
         else:
             maxval = numpy.inf
